roit-pipe/requirements.txt
minio
requests
python-dotenv
pandas